+++ src/main.rs
@@ -2,56 +2,26 @@ extern crate docopt;
 extern crate xml;
 extern crate toml;
 
-use std::fs::File;
-use std::io::prelude::*;
-use docopt::Docopt;
-
 use chunk::*;
 use file_saver::*;
 use library::*;
 
 mod chunk;
+mod config;
 mod file_saver;
 mod gobjects;
 mod library;
 mod parser;
 
-static USAGE: &'static str = "
-Usage: gir [-d <girs_dir>] [<library>]
-
-Options:
-    -d PATH            Directory for girs
-";
-
 fn main() {
-    let args = Docopt::new(USAGE).unwrap()
-        .parse().unwrap_or_else(|e| e.exit());
-
-    let toml = read_toml("Gir.toml");
-
-    let girs_dir = match args.get_str("-d") {
-        "" => toml.lookup("options.girs_dir")
-                  .unwrap_or_else(|| panic!("No options.girs_dir in config"))
-                  .as_str().unwrap(),
-        a => a
-    };
-
-    let library_name = match args.get_str("<library>") {
-        "" => toml.lookup("options.library")
-                .unwrap_or_else(|| panic!("No options.library in config"))
-                .as_str().unwrap(),
-        a => a
-    };
-
-    let objects = gobjects::parse_toml(toml.lookup("object").unwrap());
-    for object in objects.values() { println!("Objects: {:?}", object); }
+    let cfg = config::Config::new();
 
     let v = vec!["ф1", "t2", "тестs3"];
     let c = v.into_chunks();
     let tmp =  c.into_iter().save_to_file("a.txt");
 
     let mut library = Library::new();
-    library.read_file(girs_dir, library_name);
+    library.read_file(&cfg.girs_dir, &cfg.library_name);
     library.check_resolved();
     show(&library);
 }
@@ -87,26 +57,6 @@ fn show(library: &Library) {
                 continue;
             }
             println!("\tfunction {}", f.name);
-        }
-    }
-}
-
-fn read_toml(filename: &str) -> toml::Value {
-    let mut input = String::new();
-    File::open(filename).and_then(|mut f| {
-        f.read_to_string(&mut input)
-    }).unwrap();
-    let mut parser = toml::Parser::new(&input);
-    match parser.parse() {
-        Some(toml) => toml::Value::Table(toml),
-        None => {
-            for err in &parser.errors {
-                let (loline, locol) = parser.to_linecol(err.lo);
-                let (hiline, hicol) = parser.to_linecol(err.hi);
-                println!("{}:{}:{}-{}:{} error: {}",
-                         filename, loline, locol, hiline, hicol, err.desc);
-            }
-            panic!("Errors in config")
         }
     }
 }
